requirements.txt
annotated-types==0.6.0
antiorm==1.2.1
anyio==4.2.0
blinker==1.7.0
cachecontrol==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
db==0.1.1
db-dtypes==1.2.0
distro==1.9.0
exceptiongroup==1.2.0
firebase-admin==6.4.0
flask==3.0.0
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kaleido==0.2.1
MarkupSafe==2.1.3
msgpack==1.0.7
numpy==1.26.3
openai==1.7.2
packaging==23.2
pandas==2.1.4
plotly==5.18.0
proto-plus==1.23.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.1
typing-extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vanna==0.0.38
werkzeug==3.0.1
